log invalid captchas make cimment manageee page slightyl more secure

+++ admin/manage_comment.php
@@ -14,19 +14,21 @@
 	#include '../header.php';
 
 	$op = $_GET["op"];
-	$id = $_GET["id"];
-	if ($op == "a") {
-		safe_query("UPDATE `Comments` SET `Approved` = '1' WHERE `Comments`.`ID` =$id;");
-		echo "Done.";
-	}else if ($op == "d") {
-		safe_query("DELETE FROM `Comments` WHERE `Comments`.`ID` = $id AND `Approved` = '0'");
-		if (mysql_affected_rows() > 0) {			
-			echo "Record deleted.";
-		} else {
-			echo "Could not be deleted.";
+	$id = $_GET["id"]+0;//make sure it's an int
+	if ($id > 0 && is_int($id)) {
+		if ($op == "a") {
+			safe_query("UPDATE `Comments` SET `Approved` = '1' WHERE `Comments`.`ID` =$id;");
+			echo "Done.";
+		}else if ($op == "d") {
+			safe_query("DELETE FROM `Comments` WHERE `Comments`.`ID` = $id AND `Approved` = '0'");
+			if (mysql_affected_rows() > 0) {			
+				echo "Record deleted.";
+			} else {
+				echo "Could not be deleted.";
+			}
 		}
-		
 	}
+	
 	mysql_close($db);	
 
 ?>
